+++ Classes/SchedulerTasks/MailNotifierConfiguration.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\SchedulerTasks;
 
 use TYPO3\CMS\Core\Messaging\FlashMessage;
+use TYPO3\CMS\Core\Messaging\FlashMessageService;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Lang\LanguageService;
 use TYPO3\CMS\Scheduler\AdditionalFieldProviderInterface;
 use TYPO3\CMS\Scheduler\Controller\SchedulerModuleController;
@@ -17,6 +19,11 @@ use TYPO3\CMS\Scheduler\Task\AbstractTask;
  */
 class MailNotifierConfiguration implements AdditionalFieldProviderInterface
 {
+    /**
+     * @var string
+     */
+    const LABEL_PREFIX = 'LLL:EXT:seminars/Resources/Private/Language/locallang.xlf:';
+
     /**
      * Gets additional fields to render in the form to add/edit a task
      *
@@ -39,7 +46,7 @@ class MailNotifierConfiguration implements AdditionalFieldProviderInterface
         return [
             $fieldId => [
                 'code' => $fieldCode,
-                'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang.xlf:schedulerTasks.fields.page-uid',
+                'label' => self::LABEL_PREFIX . 'schedulerTasks.fields.page-uid',
             ],
         ];
     }
@@ -54,29 +61,44 @@ class MailNotifierConfiguration implements AdditionalFieldProviderInterface
      */
     public function validateAdditionalFields(array &$submittedData, SchedulerModuleController $schedulerModule): bool
     {
-        $submittedData['seminars_configurationPageUid'] = (int)$submittedData['seminars_configurationPageUid'];
-        $pageUid = $submittedData['seminars_configurationPageUid'];
+        $pageUid = (int)$submittedData['seminars_configurationPageUid'];
+        $submittedData['seminars_configurationPageUid'] = $pageUid;
         $hasPageUid = $pageUid > 0 && \Tx_Oelib_Db::existsRecordWithUid('pages', $pageUid);
         if ($hasPageUid) {
             return true;
         }
 
-        $schedulerModule->addMessage(
-            $this->getLanguageService()->sL(
-                'LLL:EXT:seminars/Resources/Private/Language/locallang.xlf:schedulerTasks.errors.page-uid'
-            ),
-            FlashMessage::ERROR
-        );
+        $message = $this->getLanguageService()->sL(self::LABEL_PREFIX . 'schedulerTasks.errors.page-uid');
+        $this->addMessage($message, FlashMessage::ERROR);
 
         return false;
     }
 
     /**
-     * @return LanguageService|null
+     * Adds a flash message.
+     *
+     * Once TYPO3 >= 9.5 is required, this class can extend `AbstractAdditionalFieldProvider`, and this method
+     * can be removed.
+     *
+     * @param string $message the flash message content
+     * @param int $severity the flash message severity
      */
-    protected function getLanguageService()
+    private function addMessage(string $message, int $severity = FlashMessage::OK)
     {
-        return $GLOBALS['LANG'] ?? null;
+        /** @var FlashMessage $flashMessage */
+        $flashMessage = GeneralUtility::makeInstance(FlashMessage::class, $message, '', $severity);
+        /** @var FlashMessageService $service */
+        $service = GeneralUtility::makeInstance(FlashMessageService::class);
+        $queue = $service->getMessageQueueByIdentifier();
+        $queue->enqueue($flashMessage);
+    }
+
+    /**
+     * @return LanguageService
+     */
+    protected function getLanguageService(): LanguageService
+    {
+        return $GLOBALS['LANG'];
     }
 
     /**